Move Advisory types to metrics _internal and export the public one
