+++ opentelemetry-api/src/opentelemetry/util/types.py
@@ -12,7 +12,6 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from dataclasses import dataclass
 from typing import Mapping, Optional, Sequence, Tuple, Union
 
 # This is the implementation of the "Any" type as specified by the specifications of OpenTelemetry data model for logs.
@@ -55,19 +54,4 @@ AttributesAsKey = Tuple[
         ],
     ],
     ...,
-]
-
-
-@dataclass
-class MetricsHistogramAdvisory:
-    explicit_bucket_boundaries: Optional[Sequence[float]] = None
-
-
-@dataclass
-class MetricsCommonAdvisory:
-    pass
-
-
-MetricsInstrumentAdvisory = Union[
-    MetricsCommonAdvisory, MetricsHistogramAdvisory
 ]
