+++ opentelemetry-sdk/tests/metrics/test_metrics.py
@@ -21,7 +21,7 @@ from typing import Iterable, Sequence
 from unittest.mock import MagicMock, Mock, patch
 
 from opentelemetry.attributes import BoundedAttributes
-from opentelemetry.metrics import NoOpMeter
+from opentelemetry.metrics import MetricsHistogramAdvisory, NoOpMeter
 from opentelemetry.sdk.environment_variables import OTEL_SDK_DISABLED
 from opentelemetry.sdk.metrics import (
     Counter,
@@ -46,7 +46,6 @@ from opentelemetry.sdk.metrics.view import SumAggregation, View
 from opentelemetry.sdk.resources import Resource
 from opentelemetry.test import TestCase
 from opentelemetry.test.concurrency_test import ConcurrencyTestBase, MockFunc
-from opentelemetry.util.types import MetricsHistogramAdvisory
 
 
 class DummyMetricReader(MetricReader):
